infrastructure/aws/cdk/app.py
#!/usr/bin/env python3
import aws_cdk as cdk
from constructs import Construct
from aws_cdk import (
    Stack,
    aws_lambda as _lambda,
    aws_iam as iam,
)

class AgenticArchitecturesStack(Stack):
    def __init__(self, scope: Construct, id: str, **kwargs) -> None:
        super().__init__(scope, id, **kwargs)

        # Create IAM role for Lambda
        bedrock_role = iam.Role(
            self, "BedrockAccessRole",
            assumed_by=iam.ServicePrincipal("lambda.amazonaws.com"),
            managed_policies=[
                iam.ManagedPolicy.from_aws_managed_policy_name(
                    "service-role/AWSLambdaBasicExecutionRole"
                )
            ]
        )

        # Add Bedrock permissions
        bedrock_role.add_to_policy(
            iam.PolicyStatement(
                actions=["bedrock:InvokeModel", "bedrock:InvokeModelWithResponseStream"],
                resources=["*"]
            )
        )

        # Create Lambda function
        agent_function = _lambda.Function(
            self, "AgentFunction",
            runtime=_lambda.Runtime.PYTHON_3_10,
            handler="index.handler",
            code=_lambda.Code.from_asset("lambda"),
            role=bedrock_role,
        )

app = cdk.App()
AgenticArchitecturesStack(app, "AgenticArchitecturesStack")
app.synth()
